+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # Change Log
 
-## Upcoming
+## 4.24.2.0
 
 - Update `code_text` to support Dark Mode with `glimmer_dark` theme, which is applied by default when app is run with OS in dark mode
 - Fix issue with `code_text` showing gibberish on Linux when Courier & Consolas fonts are not available. It will attempt to use the first Monospace font it finds now or otherwise stick to the default font.
+- Update Hello, Code Text! and Glimmer Meta-Sample to support Dark Mode
 - Refactor Hello, Custom Widget! with on_widget_disposed thread cleanup code
 - Refactor Hello, Custom Shell! to support keyboard display of emails by hitting ENTER/SPACE after selecting a table row